# db/controllers/orderController.js
import Stripe from "stripe";
import orderModel from "../models/orderModel.js";
import userModel from "../models/userModel.js";
import foodModel from "../models/foodModel.js";

const stripe = new Stripe(process.env.STRIPE_SECERT_KEY);

// placing user order for frontend
const placeOrder = async (req, res) => {
  const frontend_url = "http://localhost:5174";

  try {
    const newOrder = new orderModel({
      userId: req.body.userId,
      items: req.body.items,
      amount: req.body.amount,
      address: req.body.address,
    });
    await newOrder.save();
    await userModel.findByIdAndUpdate(req.body.userId, { cartData: {} });

    const line_items = req.body.items.map((item) => ({
      price_data: {
        currency: "usd",
        product_data: {
          name: item.name,
        },
        unit_amount: item.price * 100,
      },
      quantity: item.quantity,
    }));
    line_items.push({
      price_data: {
        currency: "usd",
        product_data: {
          name: "Delivery charges",
        },
        unit_amount: 2 * 100,
      },
      quantity: 1,
    });
    const session = await stripe.checkout.sessions.create({
      line_items: line_items,
      mode: "payment",
      success_url: `${frontend_url}/verify?success=true&orderId=${newOrder._id}`,
      cancel_url: `${frontend_url}/verify?success=false&orderId=${newOrder._id}`,
    });
    res.json({ success: true, session_url: session.url });
  } catch (error) {
    console.log(error);
    res.json({ success: false, message: "Error" });
  }
};

// verify order for frontend
const verifyOrder = async (req, res) => {
  const { orderId, success } = req.body;
  try {
    if (success === "true") {
      await orderModel.findByIdAndUpdate(orderId, { payment: true });
      res.json({ success: true, message: "Paid" });
    } else {
      await orderModel.findByIdAndDelete(orderId);
      res.json({ success: false, message: "Unpaid" });
    }
  } catch (error) {
    console.log(error);
    res.json({ success: false, message: "Error" });
  }
};

// users order for front end
const userOrders = async (req, res) => {
  try {
    const orders = await orderModel.find({ userId: req.body.userId });
    res.json({ success: true, data: orders });
  } catch (error) {
    console.log(error);
    res.json({ success: false, message: "Error" });
  }
};

// Listening orders for admin panel
const listOrders = async (req, res) => {
  try {
    const orders = await orderModel.find({});
    res.json({ success: true, data: orders });
  } catch (error) {
    console.log(error);
    res.json({ success: false, message: "Error" });
  }
};


// sales report
const sales = async (req, res) => {
  try {

    const salesdata = await orderModel.find({});
    const totalsalesreport = {
      totalsales: 0,
      numberoforders: 0,
      totalitemsSold: 0,
      averageOrderValue: 0,
      topSellingItems: {}
    }
    salesdata.forEach((value) => {
      totalsalesreport.totalsales += value.amount;
      totalsalesreport.numberoforders += 1;
      totalsalesreport.totalitemsSold += value.items.length;
      value.items.forEach((sellingitems) => {
        sellingitems._id in totalsalesreport.topSellingItems ? totalsalesreport.topSellingItems.sellingitems._id.quantity += sellingitems.quantity : totalsalesreport.topSellingItems[sellingitems._id] = [sellingitems.image, sellingitems.name, sellingitems.quantity, sellingitems.price]
      })

    });
    totalsalesreport.averageOrderValue = totalsalesreport.totalsales / totalsalesreport.numberoforders;


    const sortedObj = Object.fromEntries(
      Object.entries(totalsalesreport.topSellingItems).sort(([, a], [, b]) => b[2] - a[2])
    );

    const sortedObj2 = Object.entries(sortedObj).map(([id, [myimage, name, quantitySold, totalRevenue]]) => ({
      id,
      myimage,
      name,
      quantitySold,
      totalRevenue
    }));
    const topfive = sortedObj2.slice(0, 5)

    totalsalesreport.topSellingItems = topfive

    // console.log(totalsalesreport)

    res.json({ success: true, data: totalsalesreport });
  } catch (error) {
    console.log(error);
    res.json({ success: false, message: "Error" });
  }
};

// api for updating order status
const updateStatus = async (req, res) => {
  try {
    await orderModel.findByIdAndUpdate(req.body.orderId, {
      status: req.body.status,
    });
    res.json({ success: true, message: "Status updated" });
  } catch (error) {
    console.log(error);
    res.json({ success: false, message: "Error" });
  }
};

export { listOrders, placeOrder, updateStatus, userOrders, verifyOrder,  sales };